Longer status poll interval, switch dashboard extract func to V3 (#427)
* longer status poll interval, switch DEF to V3

* wait, not that

* 10 minutes rather than 5

* allow pasing in max attempts

* Update src/handlers/dashboard-extract/fetch-view-data.ts


* update test mock

---------
